+++ l2/agentPlannerL2/agentPlannerL2.ts
@@ -21,6 +21,7 @@ import {
   updateStatus,
 } from '/_102020_/l2/agentPlannerL2/helpers/p2Dispatch.js';
 import '/_102020_/l2/agentPlannerL2/steps/entry10/agentP2Entry.js';
+import '/_102020_/l2/agentPlannerL2/steps/workspaces20/agentP2Workspaces.js';
 
 export function createAgent(): IAgentAsync {
   return {